requirements.txt
PyPDF2
streamlit
python-dotenv
langchain
langsmith
langchain-google-genai
google-generativeai
fal-client
openai
langchain-openai
pillow
python-dotenv
requests
audio-recorder-streamlit
PyPDF2
pytesseract
pdf2image
Pillow
pymupdf
concurrent-log-handler
feedparser
langgraph
tavily-python
pandas
# --- LangChain Ecosystem (COMPATIBLE SET) ---
langchain==0.1.17
langchain-core==0.1.52
langchain-google-genai==0.0.11
langchain-openai==0.1.7
langgraph==0.0.30  
gTTS
datasets
python-dotenv

